call it stock strength and make it common function
