+++ components/Parts/BirdMouthWithPriceMA.js
@@ -12,14 +12,6 @@ const BirdMouthWithPriceMA = ({ inputTickers }) => {
   })
   const [tableLoading, setTableLoading] = useState(false)
 
-  const conditionMatches = (id, priceMAList) => {
-    const matches = !!priceMAList.find(
-      x => x.id === id && x.tickersInfo.length > 0
-    )
-
-    return matches
-  }
-
   useEffect(() => {
     ;(async () => {
       setTableSettings({
@@ -31,7 +23,7 @@ const BirdMouthWithPriceMA = ({ inputTickers }) => {
 
         const priceMAInfo = await Promise.all(
           inputTickers.map(async item => {
-            const response = await toAxios('/api/getPriceMADetails', {
+            const response = await toAxios('/api/getStockStrength', {
               ticker: item
             })
             const { data } = response
@@ -39,40 +31,9 @@ const BirdMouthWithPriceMA = ({ inputTickers }) => {
           })
         )
 
-        const tableHeader =
-          priceMAInfo.length > 0
-            ? [
-                'Ticker',
-                'Price Not Avail.',
-                'RS>0',
-                '85D High',
-                '20>50 & 50>150',
-                '20>50 & 50<150'
-              ]
-            : []
-        const tableData = priceMAInfo
-          .map(item => {
-            const twentyHigherFifty = conditionMatches(
-              '20>50',
-              item.priceMAList
-            )
-            const fiftyHigherHundredFifty = conditionMatches(
-              '50>150',
-              item.priceMAList
-            )
-            const fiftyLowerHundredFifty = conditionMatches(
-              '50<150',
-              item.priceMAList
-            )
-            return [
-              item.ticker,
-              !item.priceAvail ? 'Yes' : '',
-              item.rs > 0 ? 'Yes' : '',
-              item.latestHigherInputRange ? 'Yes' : '',
-              twentyHigherFifty && fiftyHigherHundredFifty ? 'Yes' : '',
-              twentyHigherFifty && fiftyLowerHundredFifty ? 'Yes' : ''
-            ]
-          })
+        const tableHeader = Object.keys(priceMAInfo || {})
+        const tableData = Object.values(priceMAInfo || {})
+          .map(item => (item ? 'Yes' : ''))
           .filter(item => item.find(x => x.includes('Yes')))
         const tableHeaderData = {
           tableHeader,
